+++ release/_virtual/_rollup-plugin-process-env.es6.js
@@ -1,5 +1,5 @@
 ;(function () {
-  const env = {"CANTINC_UI_VERSION":"0.1.12","CANTINC_UI_BASE_URL":""};
+  const env = {"CANTINC_UI_VERSION":"0.1.13","CANTINC_UI_BASE_URL":""};
   if (typeof process === 'undefined') {
     globalThis.process = { env: env };
   } else if (process.env) {