+++ release/ui/buttons/Buttons/Buttons.scss.es6.js
@@ -1,0 +1,7 @@
+import injector_8363bc92 from 'rollup-plugin-styles/dist/runtime/inject-css.js';
+
+var css = "form>.Buttons_root__50444128:last-child{background:var(--color-100);bottom:0;box-shadow:0 0 30px 30px var(--color-100);margin-top:48px;position:sticky}";
+var modules_fa723eb8 = {"root":"Buttons_root__50444128"};
+injector_8363bc92(css,{});
+
+export { css, modules_fa723eb8 as default };