+++ release/ui/buttons/Buttons/Buttons.scss.js
@@ -1,0 +1,16 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var injector_8363bc92 = require('rollup-plugin-styles/dist/runtime/inject-css.js');
+
+function _interopDefaultLegacy (e) { return e && typeof e === 'object' && 'default' in e ? e : { 'default': e }; }
+
+var injector_8363bc92__default = /*#__PURE__*/_interopDefaultLegacy(injector_8363bc92);
+
+var css = "form>.Buttons_root__50444128:last-child{background:var(--color-100);bottom:0;box-shadow:0 0 30px 30px var(--color-100);margin-top:48px;position:sticky}";
+var modules_fa723eb8 = {"root":"Buttons_root__50444128"};
+injector_8363bc92__default["default"](css,{});
+
+exports.css = css;
+exports["default"] = modules_fa723eb8;