+++ release/ui/icons/Icon/icons/LoupeIcon/LoupeIcon.d.ts
@@ -1,0 +1,2 @@
+import { CustomIconProps } from '../..';
+export default function LoupeIcon({ size, color, ...props }?: CustomIconProps): any;