+++ release/ui/icons/Icon/icons/LoupeIcon/index.es6.js
@@ -1,0 +1,1 @@
+export { default } from './LoupeIcon.es6.js';