+++ release/ui/icons/Icon/icons/LoupeIcon/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var LoupeIcon = require('./LoupeIcon.js');
+
+
+
+exports["default"] = LoupeIcon["default"];