+++ release/ui/interaction/Set/index.es6.js
@@ -1,1 +1,1 @@
-export { Set } from './Set.es6.js';
+export { Set, setPropsHandler } from './Set.es6.js';